+++ client/src/main/java/com/wuba/wlock/client/WReadLock.java
@@ -45,7 +45,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLock(int expireTime, int maxWaitTime) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLock(long expireTime, int maxWaitTime) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -57,7 +57,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLock(int expireTime, int maxWaitTime, int renewInterval) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLock(long expireTime, int maxWaitTime, int renewInterval) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -70,7 +70,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLock(int expireTime, int maxWaitTime, LockExpireListener lockExpireListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLock(long expireTime, int maxWaitTime, LockExpireListener lockExpireListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -83,7 +83,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLock(int expireTime, int maxWaitTime, int renewInterval, RenewListener renewListener, LockExpireListener lockExpireListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLock(long expireTime, int maxWaitTime, int renewInterval, RenewListener renewListener, LockExpireListener lockExpireListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -105,7 +105,7 @@ public class WReadLock implements WLock {
 	 * @throws ParameterIllegalException 参数校验失败异常
 	 */
 	@Override
-	public AcquireLockResult tryAcquireLock(int expireTime, HoldLockListener holdLockListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLock(long expireTime, HoldLockListener holdLockListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -146,7 +146,7 @@ public class WReadLock implements WLock {
 
 
 	@Override
-	public AcquireLockResult tryAcquireLockUnblocked(int expireTime) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLockUnblocked(long expireTime) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -157,7 +157,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLockUnblocked(int expireTime, int renewInterval) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLockUnblocked(long expireTime, int renewInterval) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -176,7 +176,7 @@ public class WReadLock implements WLock {
 	 * @throws ParameterIllegalException 参数校验失败异常
 	 */
 	@Override
-	public AcquireLockResult tryAcquireLockUnblocked(int expireTime, HoldLockListener holdLockListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLockUnblocked(long expireTime, HoldLockListener holdLockListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -188,7 +188,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLockUnblocked(int expireTime, LockExpireListener lockExpireListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLockUnblocked(long expireTime, LockExpireListener lockExpireListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -200,7 +200,7 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public AcquireLockResult tryAcquireLockUnblocked(int expireTime, int renewInterval, RenewListener renewListener, LockExpireListener lockExpireListener) throws ParameterIllegalException {
+	public AcquireLockResult tryAcquireLockUnblocked(long expireTime, int renewInterval, RenewListener renewListener, LockExpireListener lockExpireListener) throws ParameterIllegalException {
 		LockOption lockOption = new LockOption();
 		lockOption.setLockType((byte) LockTypeEnum.readWriteReentrantLock.getValue());
 		lockOption.setLockOpcode((byte) ReadWriteLockTypeEnum.Read.getOpcode());
@@ -237,20 +237,20 @@ public class WReadLock implements WLock {
 	}
 
 	@Override
-	public LockResult renewLock(int expireTime) throws ParameterIllegalException {
+	public LockResult renewLock(long expireTime) throws ParameterIllegalException {
 		return renewLock(-1, expireTime);
 	}
 
-	private LockResult renewLock(long lockversion, int expireTime) throws ParameterIllegalException {
+	private LockResult renewLock(long lockversion, long expireTime) throws ParameterIllegalException {
 		long threadID = Thread.currentThread().getId();
 		return renewLock(lockversion, threadID, expireTime);
 	}
 
-	private LockResult getRenewLockResult(long lockversion,long ownerThreadID, int expireTime) throws ParameterIllegalException {
+	private LockResult getRenewLockResult(long lockversion,long ownerThreadID, long expireTime) throws ParameterIllegalException {
 		return this.wlockClient.getLockService().renewLock(lockkey, lockversion, expireTime, ownerThreadID, LockTypeEnum.readWriteReentrantLock.getValue(), ReadWriteLockTypeEnum.Read.getOpcode());
 	}
 
-	private LockResult renewLock(long lockversion, long ownerThreadID, int expireTime) throws ParameterIllegalException {
+	private LockResult renewLock(long lockversion, long ownerThreadID, long expireTime) throws ParameterIllegalException {
 		if (LockPolicy.Process == lockPolicy) {
 			ownerThreadID = PROCESS_LOCK_THREAD_ID;
 			synchronized (lockkey.intern()) {
@@ -260,7 +260,7 @@ public class WReadLock implements WLock {
 		return getRenewLockResult(lockversion, ownerThreadID, expireTime);
 	}
 
-	private void dealExpireTime(LockOption lockOption , int expireTime){
+	private void dealExpireTime(LockOption lockOption , long expireTime){
 		if (expireTime > Factor.LOCK_MAX_EXPIRETIME) {
 			lockOption.setExpireTime(Factor.LOCK_MAX_EXPIRETIME);
 			lockOption.setRenewInterval(Factor.HOLD_LOCK_RENEWINTERVAL);
